requirements.txt
torch
nibabel
monai
PySide6
torchio
Pillow